refactor: make the component just a hook
