+++ packages/api-client/src/store/import-spec.ts
@@ -29,7 +29,6 @@ export function importSpecFileFactory({
     options: ImportSpecFileArgs = {},
   ) => {
     const spec = toRaw(_spec)
-    console.log('import spec', spec)
     const workspaceEntities = await importSpecToWorkspace(spec, options)
 
     if (workspaceEntities.error) {
